move ResyncLoop to utils.go
