+++ pkg/plugins/cloudfoundry/endpoint_controller.go
@@ -16,33 +16,6 @@ import (
 	"github.com/solo-io/gloo/pkg/log"
 )
 
-func ResyncLoop(ctx context.Context, stop <-chan struct{}, resync func(), resyncDuration time.Duration) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	ticker := time.NewTicker(resyncDuration)
-	defer ticker.Stop()
-
-	if stop != nil {
-		go func(cancel context.CancelFunc) {
-			<-stop
-			cancel()
-		}(cancel)
-	}
-
-	// first time resync
-	resync()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			resync()
-		}
-	}
-}
-
 func createEndpointDiscovery(ctx context.Context, opts bootstrap.Options) (endpointdiscovery.Interface, error) {
 	// kubeConfig := opts.KubeOptions.KubeConfig
 	// masterUrl := opts.KubeOptions.MasterURL
